impl of Planner: moved codes of main func to one for unit test.
